fix remote memory bug + UI fix

+++ Just Diary/Services/DatabaseDataManager.swift
@@ -10,6 +10,7 @@ import CoreData
 
 class DatabaseDataManager {
     
+    private var memoriesObjects: [MemoryObject] = []
     private let container = NSPersistentContainer(name: "MainData")
     private var managedContext: NSManagedObjectContext?
     
@@ -21,9 +22,10 @@ class DatabaseDataManager {
         do {
             let result = try managedContext.fetch(memoryFetch)
             var value: [Memory] = []
-            for res in result {
-                let temp = Memory(fromMemoreObject: res)
+            for memoryObject in result {
+                let temp = Memory(fromMemoreObject: memoryObject)
                 value.append(temp)
+                self.memoriesObjects.append(memoryObject)
             }
             return value
         } catch let error as NSError {
@@ -34,15 +36,16 @@ class DatabaseDataManager {
     
     func saveData(memory: Memory) {
         guard let managedContext = self.managedContext else { return }
-        _ = MemoryObject(byMemory: memory, andContext: managedContext)
+        let memoryObject = MemoryObject(byMemory: memory, andContext: managedContext)
+        self.memoriesObjects.append(memoryObject)
         self.saveContext()
     }
     
-    func delete(memory: Memory) {
+    func deleteMemory(byId id: UUID) {
         guard let managedContext = self.managedContext else { return }
-        let memoryObject = MemoryObject(byMemory: memory, andContext: managedContext)
+        guard let memoryObject = self.memoriesObjects.first(where: {$0.id == id}) else { return }
         managedContext.delete(memoryObject)
-        self.saveContext()
+        self.saveContext(managedContext)
     }
     
     func update(memory: Memory) {
@@ -54,11 +57,13 @@ class DatabaseDataManager {
     
     // MARK: - Private func
     
-    private func saveContext() {
-        guard let managedContext = self.managedContext else { return }
+    private func saveContext(_ context: NSManagedObjectContext? = nil) {
+        guard let managedContext = context ?? self.managedContext else { return }
+//        guard let managedContext = self.managedContext else { return }
         guard managedContext.hasChanges else { return }
         do {
             try managedContext.save()
+            print("contex saved")
         } catch let error as NSError {
             print("Unresolved error \(error), \(error.userInfo)")
         }
